// src/types/dialogs.ts
export enum DialogType {
	CASHBACK_INFO = "CASHBACK_INFO",
	CREATE_PROVIDER_STEPS = "CREATE_PROVIDER_STEPS",
	UPLOAD_PROVIDER_FILES = "UPLOAD_PROVIDER_FILES",
	CONTINUE_WITH_EXISTING_DATA = "CONTINUE_WITH_EXISTING_DATA",
	PROFILE_CREATED = "PROFILE_CREATED",
	REVIEW_CAROUSEL = "REVIEW_CAROUSEL",
	VERIFY_EMAIL = "VERIFY_EMAIL",
	PROVIDERS_CONTACT_CONFIRMATION = "PROVIDERS_CONTACT_CONFIRMATION",
	PROVIDERS_REVIEW_CONFIRMATION = "PROVIDERS_REVIEW_CONFIRMATION",
}

export type DialogState = {
	[key in DialogType]?: {
		isOpen: boolean;
		data?: unknown;
	};
};